+++ app/Http/Controllers/ContactController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\ContactUsRequest;
-use Illuminate\Mail\Message;
+use App\Services\ContactUsMailer;
 
 
 class ContactController extends Controller
@@ -11,27 +11,12 @@ class ContactController extends Controller
     public function index(){
         return view('contact.contact');
     }
-    public function send(ContactUsRequest $request){
+    public function send(ContactUsRequest $request , ContactUsMailer $mailer){
 
         $data = $request->validated();
+        $mailer->send($data);
 
-         \Mail::send(
-             "emails.contactUs",
-             [
-                 'name' => $data['name'],
-                 'email'=> $data['email'],
-                 'phone' => $data['phone'],
-                 'country' => $data['country'],
-                 'region' => $data['region'],
-                 'messageText'=> $data['message']
-
-             ], 
-             function (Message $message) use ($data){
-                 $message->subject('Contact Us requested from ' . $data['email']);
-                 $message->from($data['email'], $data['name']);
-                 $message->to('e-commerce@clokids.app');
-                 
-         });
+        
        return  redirect()->route('contact')->withInput($data);
     }
 }
